NOTE: "pardon the dust"
master currently contains files that are loaded but incomplete / not
designed / not usually appropriate for "master".

This means that master is not quite in "working state" since it was just
the result of a huge merge in 1dcf7332cfb1179b89cd46a236fa104d17500d0e.

This means that:
- lint is broken, so this commit forcibly disables it in these files.
- some files might not work as expected

Rolling back would roll back other changes to, so it's staying this way.